# apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the CyVerse Discovery Environment: app discovery and launching, analysis management, and iRODS data access.\n\nTo authenticate: open the Authorize dialog and enter your username and password under BasicAuth,\nrun POST /login, then copy the access_token from the response into the BearerAuth value\n(prefixed with \"Bearer \") to call the other endpoints.",
        "title": "formation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/app/launch/{system_id}/{app_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits an analysis. The body is optional; missing fields get defaults: a generated\nanalysis name, an output directory under the user's home, the email from the JWT,\ndebug=false, and notify=true. Swagger UI placeholder values are stripped.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apps"
                ],
                "summary": "Launch an app",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution system id (e.g. de)",
                        "name": "system_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "App UUID",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "iRODS zone for the output directory (defaults to the configured zone)",
                        "name": "output_zone",
                        "in": "query"
                    },
                    {
                        "description": "Analysis submission; all fields optional",
                        "name": "submission",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.LaunchSubmission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "analysis_id, name, status, plus url for VICE analyses",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid app ID or request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/apps/analyses/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists analyses filtered by status (default Running). Pass status with an empty value to list all analyses.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyses"
                ],
                "summary": "List the user's analyses",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Running",
                        "description": "Analysis status filter; explicitly empty lists all",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "analyses with analysis_id, name, app_id, system_id, and status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/apps/analyses/{analysis_id}/control": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "extend_time extends the analysis's time limit; save_and_exit terminates it after saving outputs; exit terminates it without saving.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyses"
                ],
                "summary": "Control a running VICE analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analysis UUID",
                        "name": "analysis_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "extend_time",
                            "save_and_exit",
                            "exit"
                        ],
                        "type": "string",
                        "description": "Operation to perform",
                        "name": "operation",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation result; includes the echoed operation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid operation or analysis ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/apps/analyses/{analysis_id}/details": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyses"
                ],
                "summary": "Get an analysis's full details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analysis UUID",
                        "name": "analysis_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The analysis record as returned by the apps service",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid analysis ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Apps service error (including unknown analyses)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/apps/analyses/{analysis_id}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyses"
                ],
                "summary": "Get an analysis's status and VICE URL readiness",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analysis UUID",
                        "name": "analysis_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "analysis_id, status, url_ready, plus url and url_check_details when a subdomain exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid analysis ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Apps service error (including unknown analyses)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.LaunchSubmission": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "debug": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notify": {
                    "type": "boolean"
                },
                "output_dir": {
                    "type": "string"
                },
                "requirements": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "A Keycloak JWT prefixed with \"Bearer \". Obtain one via POST /login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}